The bounding box will be recalculated each time it is modified by any Set..() method

+++ includes/visualization/ogl_chart_geometry_c.h
@@ -4,7 +4,7 @@
 #include "chart_types.h"
 
 //! Defines which datatype to use for position calculations
-using PositionType_TP = int;
+using PositionType_TP = float;
 
 //! Defines the position and size of the chart inside the opengl context
 //!
